+++ src/components/Resources/TN/index.js
@@ -1,0 +1,1 @@
+export * from './TN';
