--- src/components/Resources/TNTW/index.js
@@ -1,1 +1,0 @@
-export * from './TNTW';
