+++ src/components/index.js
@@ -5,8 +5,8 @@ export { default as MenuButtons } from './Notes/MenuButtons';
 export { default as DropdownMenu } from './Notes/DropdownMenu';
 export { default as Workspace } from './Workspace/Workspace';
 export { default as Bible } from './Resources/Bible/Bible';
-export { default as TN } from './Resources/TNTW/TN';
-export { default as TNContent } from './Resources/TNTW/TNContent';
-export { default as TNList } from './Resources/TNTW/TNList';
+export { default as TNotes } from './Resources/TN/TNotes';
+export { default as TNContent } from './Resources/TN/TNContent';
+export { default as TNList } from './Resources/TN/TNList';
 export { default as TQuestion } from './Resources/TQ/TQuestion';
-export { default as TQuestionsList } from './Resources/TQ/TQuestionsList';
+export { default as TQuestions } from './Resources/TQ/TQuestions';
